Add app to automatically update dependency versions
The update-dependencies .NET app must be invoked from the root of the
repo.  A update-dependencies.ps1 script is also provided which installs
the .NET CLI and then compiles and invokes the app; this script can be
invoked from anywhere.

The plan is for Maestro to invoke a VSTS build definition whenever a new
CLI build is available, which will invoke the update-dependencies.ps1
script.

Also add a .gitignore file.